requirements.txt
twitch-python==0.0.20
TTS==0.7.1
configparser==5.2.0
